mv runtime config bindings to new file local_model.rs
